# src/main/java/com/collabera/library/dto/LibraryResponse.java
package com.collabera.library.dto;

import java.time.LocalDate;

public class LibraryResponse {

    private String bookTitle;
    private String borrowerName;
    private LocalDate localDate;
    private String message;

    public String getBookTitle() {
        return bookTitle;
    }

    public void setBookTitle(String bookTitle) {
        this.bookTitle = bookTitle;
    }

    public String getBorrowerName() {
        return borrowerName;
    }

    public void setBorrowerName(String borrowerName) {
        this.borrowerName = borrowerName;
    }

    public LocalDate getLocalDate() {
        return localDate;
    }

    public void setLocalDate(LocalDate localDate) {
        this.localDate = localDate;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }
}