# swagger/library-api.yaml
openapi: 3.1.0
info:
  title: Collabera Library API
  description: API for managing Library tasks
  version: v1
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/v1/library/return:
    post:
      tags:
        - Return a book
      summary: Create a return book
      operationId: returnBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowBookRequest'
        required: true
      responses:
        '201':
          description: Successfully borrowed a book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryResponse'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: string
                example: Internal server error
  /api/v1/library/borrow:
    post:
      tags:
        - Borrow a book
      summary: Create a borrow book
      operationId: borrowBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowBookRequest'
        required: true
      responses:
        '201':
          description: Successfully borrowed a book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibraryResponse'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: string
                example: Internal server error
  /api/v1/borrower:
    post:
      tags:
        - Create Borrower
      summary: Create a borrower
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BorrowerRequest'
        required: true
      responses:
        '201':
          description: Successfully created borrower
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BorrowerResponse'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: string
                example: Internal server error
  /api/v1/books:
    get:
      tags:
        - Get All Books
      summary: Retrieve All of Books
      operationId: getAll
      responses:
        '200':
          description: Successful retrieval
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookResponse'
    post:
      tags:
        - Register Book
      summary: Register a Book
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookRequest'
        required: true
      responses:
        '201':
          description: Successfully created Book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResponse'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: string
                example: Internal server error
components:
  schemas:
    BorrowBookRequest:
      type: object
      properties:
        bookID:
          type: integer
          format: int64
        borrowerId:
          type: integer
          format: int64
      required:
        - bookID
        - borrowerId
    LibraryResponse:
      type: object
      properties:
        bookTitle:
          type: string
        borrowerName:
          type: string
        localDate:
          type: string
          format: date
        message:
          type: string
    BorrowerRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 2
        email:
          type: string
          format: email
          minLength: 1
      required:
        - email
        - name
    BorrowerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
    BookRequest:
      type: object
      properties:
        isbn:
          type: string
          maxLength: 50
          minLength: 2
        title:
          type: string
          maxLength: 50
          minLength: 2
        author:
          type: string
          maxLength: 50
          minLength: 2
      required:
        - author
        - isbn
        - title
    BookResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        isbn:
          type: string
        title:
          type: string
        author:
          type: string
        availability:
          type: boolean